auth0 more functional
